Allow options to be set with environment variables in lambda
Prior to this commit, the way to customize your installation of
echo-sonos was to make a copy of the options.example.js and hand
edit it for your setup.

After this commit, the way to customize your installation of
echo-sonos can be that you use the same options.js as everyone
else but you set the items you want to change as environment
variables when you upload to lamdba.