fixed bug whereby you could not set any environment variable to logical false via the lambda console.
The issue is that the code was setting the environment variable to
string value "true" and string value "false" which are both truthy.

only users trying to use node-sonos via http noticed this bug because
there are only two boolean env vars: one is useSQS and if you are ok
with true for this, you don't care what the value of useHttps is.